Avoid to update inlined relation column to NULL when deleting entities
When deleting entities, cubicweb run a rql DELETE on all relations to trigger hooks.
For an inlined relation this also mean set the column to NULL. This operation
may fail if there's additional constraints on the column.

Also this is a weird and useless behavior since deleting the entity row will by
definition delete the relation.

We still doesn't handle the case where both subject and object are going to be
deleted because rows need to be deleted in a particular order that cubicweb
doesn't handle.

Add a test checking UPDATE does not occur but hooks are correctly called.

Closes #17236690

--HG--
branch : 3.26